refactor: added enum to packages/api
