+++ packages/api/src/functions/checkout.ts
@@ -1,13 +1,14 @@
 import { HttpRequest, HttpResponseInit, InvocationContext } from "@azure/functions";
 import { getConfig, initializeDatabaseConfiguration } from "../config";
-import { ReservationRequest } from "../models/reservation-request";
+import { ReservationRequest } from "../interface/models";
 import { getListingById } from "../models/listing";
 import {
   findReservationsByListingIdAndDateRange,
   saveReservation,
   updateReservationStatus,
 } from "../models/reservation";
-import { Listing } from "../models/listing.schema";
+import { Listing } from "../interface/models";
+import { ReservationStatus } from "../types/models";
 
 // POST: Checkout
 export async function postCheckout(request: HttpRequest, context: InvocationContext): Promise<HttpResponseInit> {
@@ -114,7 +115,7 @@ export async function postCheckout(request: HttpRequest, context: InvocationCont
       guests,
       from,
       to,
-      status: "pending" as const,
+      status: ReservationStatus.Pending as const,
       createdAt: now,
     };
 
@@ -157,7 +158,7 @@ export async function postCheckout(request: HttpRequest, context: InvocationCont
     } catch (error: unknown) {
       const err = error as Error;
       context.error(`Error creating checkout session: ${err.message}`);
-      await updateReservationStatus(reservationRecord.id, "cancelled");
+      await updateReservationStatus(reservationRecord.id, ReservationStatus.Cancelled);
       throw error;
     }
   } catch (error: unknown) {
