+++ packages/api/src/functions/payments.ts
@@ -3,7 +3,8 @@ import { initializeDatabaseConfiguration } from "../config";
 import { findPaymentById, findPaymentsByUserId, savePayment, isValidPayment } from "../models/payment";
 import { findUserById } from "../models/user";
 import { updateReservationStatus } from "../models/reservation";
-import { Payment } from "../models/payment.schema";
+import { Payment } from "../interface/models";
+import { ReservationStatus } from "../types/models";
 
 // GET: Get Payment by Id
 export async function getPaymentById(request: HttpRequest, context: InvocationContext): Promise<HttpResponseInit> {
@@ -127,7 +128,7 @@ export async function postPayment(request: HttpRequest, context: InvocationConte
       };
     }
 
-    const reservationRecord = await updateReservationStatus(payment.reservationId, "active");
+    const reservationRecord = await updateReservationStatus(payment.reservationId, ReservationStatus.Active);
     if (!reservationRecord) {
       context.error(`Error payment received for unknown reservation id: ${payment.reservationId}`);
       return {
