+++ packages/api/src/functions/reservations.ts
@@ -1,7 +1,8 @@
 import { HttpRequest, HttpResponseInit, InvocationContext } from "@azure/functions";
 import { initializeDatabaseConfiguration } from "../config";
 import { findReservationById, findReservationsByUserId, updateReservationStatus } from "../models/reservation";
-import { Reservation } from "../models/reservation.schema";
+import { Reservation } from "../interface/models";
+import { ReservationStatus } from "../types/models";
 
 // GET: Get Reservation by Id
 export async function getReservationById(request: HttpRequest, context: InvocationContext): Promise<HttpResponseInit> {
@@ -115,7 +116,7 @@ export async function patchReservationById(
         error: "Reservation status is missing",
       },
     };
-  } else if (status !== "active" && status !== "cancelled") {
+  } else if (status !== ReservationStatus.Active && status !== ReservationStatus.Cancelled) {
     return {
       status: 400,
       jsonBody: {
