+++ packages/api/src/models/favorite.ts
@@ -1,7 +1,8 @@
 import { pgQuery } from "../config/pgclient";
-import FavoriteModel, { Favorite } from "./favorite.schema";
+import FavoriteModel from "./favorite.schema";
+import { Favorite } from "../interface/models";
 import { listingMapper } from "./listing";
-import { Listing } from "./listing.schema";
+import { Listing } from "../interface/models";
 
 export async function saveFavorite(fav: Favorite): Promise<Favorite | null> {
   const recordFavorite = await findFavorite(fav);
